rag/requirements.txt
langchain_community==0.2.12
langchain_core==0.2.33
langchain_openai==0.1.22
langchain_pinecone==0.1.3
nltk==3.8.1
pandas==2.2.2
pinecone==5.0.1
python-dotenv==1.0.1
